[refactor] Backend code organization
- extract `api.js`;
- ensure API returns comment/reply author and creator;
- move `utils.js` used only by backend tests into
`static/tests/backend/` folder;
