+++ ep.json
@@ -17,10 +17,10 @@
         "aceInitialized": "ep_comments_page/static/js/index"
       },
       "hooks": {
+        "expressCreateServer": "ep_comments_page/api",
         "padRemove": "ep_comments_page/index",
         "padCopy": "ep_comments_page/index",
         "socketio": "ep_comments_page/index",
-        "expressCreateServer": "ep_comments_page/index",
         "collectContentPre": "ep_comments_page/static/js/shared",
         "eejsBlock_scripts": "ep_comments_page/index",
         "exportHtmlAdditionalTagsWithData": "ep_comments_page/exportHTML",
