+++ static/tests/backend/specs/api/exportHTML.js
@@ -2,7 +2,7 @@ var supertest = require('ep_etherpad-lite/node_modules/supertest'),
            fs = require('fs'),
          path = require('path'),
       request = require('request'),
-        utils = require('../../../utils'),
+        utils = require('../../utils'),
        apiKey = utils.apiKey,
     codeToBe0 = utils.codeToBe0,
           api = utils.api,
